2rd: filter out deprecated ut
