+++ Makefile.common
@@ -67,7 +67,7 @@ endif
 
 LINUX     := "Linux"
 MAC       := "Darwin"
-PACKAGE_LIST  := go list ./...| grep -vE "docs|proto|library/util|library/client|library/copywriting|micro-api|file-server|micro-metadb/service|micro-cluster/datatransfer|micro-cluster/cluster|micro-cluster/service/cluster/adapt|micro-cluster/service/tenant/adapt|^github.com/pingcap-inc/tiem/micro-cluster/service$$|clusterpb|dbpb|adapt|test/mock|micro-cluster/parametergroup|micro-cluster/cluster/parameter"
+PACKAGE_LIST  := go list ./...| grep -vE "docs|proto|library/util|library/client|library/copywriting|micro-api|file-server|micro-metadb/service|micro-cluster/datatransfer|micro-cluster/cluster|micro-cluster/service/cluster/adapt|micro-cluster/service/tenant/adapt|^github.com/pingcap-inc/tiem/micro-cluster/service$$|clusterpb|dbpb|adapt|test/mock|micro-cluster/parametergroup|micro-cluster/cluster/parameter|deprecated"
 PACKAGES  ?= $$($(PACKAGE_LIST))
 #PACKAGE_DIRECTORIES := $(PACKAGE_LIST) | sed 's|github.com/pingcap/$(PROJECT)||'
 PACKAGE_DIRECTORIES := $(PACKAGE_LIST)
