+++ library/secondparty/lib_api_deprecated.go
@@ -15,11 +15,11 @@
  ******************************************************************************/
 
 /*******************************************************************************
- * @File: lib_api
- * @Description:
- * @Author: shenhaibo@pingcap.com
+ * @File: lib_api_deprecated.go
+ * @Description: tidb component uses api to update parameters
+ * @Author: jiangxunyu@pingcap.com
  * @Version: 1.0.0
- * @Date: 2021/12/15
+ * @Date: 2021/12/2 13:51
 *******************************************************************************/
 
 package secondparty
@@ -30,18 +30,15 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/pingcap-inc/tiem/library/framework"
-	"github.com/pingcap-inc/tiem/library/spec"
 	util "github.com/pingcap-inc/tiem/library/util/http"
+
+	"github.com/pingcap-inc/tiem/library/spec"
+
+	"github.com/pingcap-inc/tiem/library/framework"
 )
 
-const (
-	PdApiUrl   = "/pd/api/v1/config"
-	TiKVApiUrl = "/config"
-	TiDBApiUrl = "/settings"
-)
-
-func (manager *SecondPartyManager) ApiEditConfig(ctx context.Context, apiEditConfigReq ApiEditConfigReq) (bool, error) {
+// Deprecated
+func (secondMicro *SecondMicro) ApiEditConfig(ctx context.Context, apiEditConfigReq ApiEditConfigReq) (bool, error) {
 	framework.LogWithContext(ctx).Infof("micro srv api edit config, api req: %v", apiEditConfigReq)
 	switch apiEditConfigReq.TiDBClusterComponent {
 	case spec.TiDBClusterComponent_TiDB:
