+++ library/secondparty/lib_sql_test.go
@@ -6,11 +6,12 @@
  *                                                                            *
  * http://www.apache.org/licenses/LICENSE-2.0                                 *
  *                                                                            *
- *  Unless required by applicable law or agreed to in writing, software       *
- *  distributed under the License is distributed on an "AS IS" BASIS,         *
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  *
- *  See the License for the specific language governing permissions and       *
- *  limitations under the License.                                            *
+ * Unless required by applicable law or agreed to in writing, software        *
+ * distributed under the License is distributed on an "AS IS" BASIS,          *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+ * See the License for the specific language governing permissions and        *
+ * limitations under the License.                                             *
+ *                                                                            *
  ******************************************************************************/
 
 /*******************************************************************************
@@ -18,7 +19,7 @@
  * @Description:
  * @Author: shenhaibo@pingcap.com
  * @Version: 1.0.0
- * @Date: 2021/12/6
+ * @Date: 2021/12/15
 *******************************************************************************/
 
 package secondparty
@@ -33,122 +34,17 @@ import (
 	"github.com/pingcap-inc/tiem/library/spec"
 )
 
-var secondMicro3 *SecondMicro
 var secondPartyManager3 *SecondPartyManager
 var dbConnParam3 DbConnParam
 var req ClusterEditConfigReq
 var manager *SecondPartyManager
 
 func init() {
-	secondMicro3 = &SecondMicro{}
-	secondPartyManager2 = &SecondPartyManager{}
+	secondPartyManager3 = &SecondPartyManager{}
 	dbConnParam3 = DbConnParam{
 		Username: "root",
 		IP:       "127.0.0.1",
 		Port:     "4000",
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v1(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				TiDBClusterComponent: spec.TiDBClusterComponent_TiKV,
-				ConfigKey:            "split.qps-threshold",
-				ConfigValue:          "1000",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil {
-		t.Error("err nil")
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v2(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				InstanceAddr:         "127.0.0.1:10020",
-				TiDBClusterComponent: spec.TiDBClusterComponent_TiKV,
-				ConfigKey:            "split.qps-threshold",
-				ConfigValue:          "3000",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil {
-		t.Error("err nil")
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v3(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				TiDBClusterComponent: spec.TiDBClusterComponent_TiKV,
-				ConfigKey:            "log-level",
-				ConfigValue:          "'warn'",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil {
-		t.Error("err nil")
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v4(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				TiDBClusterComponent: spec.TiDBClusterComponent_PD,
-				ConfigKey:            "log.level",
-				ConfigValue:          "'info'",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil {
-		t.Error("err nil")
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v5(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				TiDBClusterComponent: spec.TiDBClusterComponent_TiDB,
-				ConfigKey:            "tidb_slow_log_threshold",
-				ConfigValue:          "200",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil {
-		t.Error("err nil")
-	}
-}
-
-func TestSecondMicro_EditClusterConfig_Deprecated_v6(t *testing.T) {
-	req = ClusterEditConfigReq{
-		DbConnParameter: dbConnParam3,
-		ComponentConfigs: []ClusterComponentConfig{
-			{
-				TiDBClusterComponent: spec.TiDBClusterComponent_TiFlash,
-				ConfigKey:            "tidb_slow_log_threshold",
-				ConfigValue:          "200",
-			},
-		},
-	}
-	err := secondMicro3.EditClusterConfig(context.TODO(), req, 0)
-	if err == nil || !strings.Contains(err.Error(), "not support") {
-		t.Errorf("err nil or err(%s) not contain not support", err.Error())
 	}
 }
